Move web3 to app folder
