+++ web/src/index.tsx
@@ -5,7 +5,7 @@ import App from './App';
 import {store} from './app/store';
 import {Provider} from 'react-redux';
 import * as serviceWorker from './serviceWorker';
-import {dispatchEthereumEvents} from "./features/web3/events";
+import {dispatchEthereumEvents} from "./features/wallet/events";
 
 const container = document.getElementById('root')!;
 const root = ReactDOMClient.createRoot(container);
